Fix: Setting parsed serialization failed tests
- Setting the parsed serialization will result in hasChanged being
set to true, which will cause the entry to be marked as changed. This
caused differences when the original entry was not serlized.

The fix has added a check of original entry to see if it has
parsed serialization, and if it does, it will be set to the new
entry.

Link to DD2480-Group1/jabref#15